using uuid for user to better integrate with oidc services like globus. eliminated the need to Identity model. added tests for user setup.
