+++ app/views.py
@@ -1,6 +1,8 @@
-from django.contrib.auth import get_user_model
-from django.http import HttpResponse, HttpResponseRedirect, Http404
-from django.views.generic import FormView
+from django.conf import settings
+from django.contrib.auth import login, logout, get_user_model
+from django.http import HttpRequest, HttpResponse, HttpResponseRedirect, HttpResponseBadRequest, Http404
+from django.shortcuts import redirect
+from django.views.generic.edit import FormView
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -8,7 +10,7 @@ from rest_framework.generics import ListAPIView, RetrieveAPIView
 from rest_framework.permissions import IsAdminUser, IsAuthenticated
 from rest_framework.authtoken.models import Token
 from .serializers import UserSerializer
-from .forms import UpdateSSHPublicKeysForm
+from .forms import UpdateSSHPublicKeysForm, CompleteLoginForm
 
 User = get_user_model()
 
@@ -28,7 +30,6 @@ class UserDetailView(RetrieveAPIView):
 
 class UserSelfDetailView(RetrieveAPIView):
     serializer_class = UserSerializer
-    # TODO allow self permissions
     permission_classes = [IsAuthenticated]
 
     def get_object(self):
@@ -76,9 +77,9 @@ class UpdateSSHPublicKeysView(FormView):
     success_url = "/"
     
     def get_initial(self):
-        data = super().get_initial()
-        data["ssh_public_keys"] = self.request.user.ssh_public_keys
-        return data
+        return {
+            "ssh_public_keys": self.request.user.ssh_public_keys
+        }
 
     def form_valid(self, form) -> HttpResponse:
         cleaned_data = form.cleaned_data
@@ -86,3 +87,46 @@ class UpdateSSHPublicKeysView(FormView):
         user.ssh_public_keys = cleaned_data["ssh_public_keys"]
         user.save()
         return HttpResponseRedirect("/")
+
+
+class CompleteLoginView(FormView):
+
+    form_class = CompleteLoginForm
+    template_name = None
+
+    def get(self, request: HttpRequest) -> HttpResponse:
+        try:
+            user_info = request.session["oidc_auth_user_info"]
+        except KeyError:
+            return HttpResponseBadRequest("missing oidc_auth_user_info session key")
+
+        try:
+            sub = user_info["sub"]
+        except KeyError:
+            return HttpResponseBadRequest("missing oidc user info subject from session data")
+
+        try:
+            user = User.objects.get(id=sub)
+        except User.DoesNotExist:
+            return super().get(request)
+
+        return self.login_and_redirect(user)
+
+    def form_valid(self, form):
+        cleaned_data = form.cleaned_data
+
+        user_info = self.request.session["oidc_auth_user_info"]
+
+        user = User.objects.create(
+            id=user_info["sub"],
+            username=cleaned_data["username"],
+            name=user_info.get("name"),
+            email=user_info.get("email"),
+            organization=user_info.get("organization"),
+        )
+
+        return self.login_and_redirect(user)
+    
+    def login_and_redirect(self, user):
+        login(self.request, user)
+        return redirect(self.request.session.get("oidc_auth_next", settings.LOGIN_REDIRECT_URL))
